Rename qemu folder to libvirt and fix imports
This commit contains only the rename and import fix changes.
No other code is changed.
