+++ lisa/mixin_modules.py
@@ -21,11 +21,11 @@ import lisa.sut_orchestrator.azure.transformers  # noqa: F401
 import lisa.sut_orchestrator.ready  # noqa: F401
 
 if platform.system() == "Linux":
-    import lisa.sut_orchestrator.qemu.context  # noqa: F401
-    import lisa.sut_orchestrator.qemu.platform  # noqa: F401
-    import lisa.sut_orchestrator.qemu.schema  # noqa: F401
-    import lisa.sut_orchestrator.qemu.qemu_platform  # noqa: F401
-    import lisa.sut_orchestrator.qemu.ch_platform  # noqa: F401
+    import lisa.sut_orchestrator.libvirt.context  # noqa: F401
+    import lisa.sut_orchestrator.libvirt.platform  # noqa: F401
+    import lisa.sut_orchestrator.libvirt.schema  # noqa: F401
+    import lisa.sut_orchestrator.libvirt.qemu_platform  # noqa: F401
+    import lisa.sut_orchestrator.libvirt.ch_platform  # noqa: F401
 
 import lisa.transformers.dump_variables  # noqa: F401
 import lisa.transformers.kernel_source_installer  # noqa: F401
