+++ lisa/sut_orchestrator/libvirt/ch_platform.py
@@ -10,8 +10,8 @@ from lisa import schema
 from lisa.environment import Environment
 from lisa.feature import Feature
 from lisa.node import Node
-from lisa.sut_orchestrator.qemu.context import NodeContext, get_node_context
-from lisa.sut_orchestrator.qemu.platform import BaseLibvirtPlatform
+from lisa.sut_orchestrator.libvirt.context import NodeContext, get_node_context
+from lisa.sut_orchestrator.libvirt.platform import BaseLibvirtPlatform
 from lisa.tools import QemuImg
 from lisa.util.logger import Logger
