+++ lisa/sut_orchestrator/libvirt/console_logger.py
