+++ lisa/sut_orchestrator/libvirt/platform.py
